Get a mapping table of CALIBER phenotypes/categories

Includes a category order and colour for plotting. These were manually assigned based on those used in the PheWAS::PheWAS package (see PheWAS::pheinfo). The original CALIBER phenotype to categories mapping file was obtained from CALIBER's github repo.

Usage

get_caliber_categories_mapping()

Value

A data frame

Examples

get_caliber_categories_mapping()
#> # A tibble: 308 × 4
#>    phenotype                                  category           groupnum colour
#>    <chr>                                      <chr>                 <int> <chr> 
#>  1 Bacterial Diseases (excl TB)               Infectious Diseas…        1 blue  
#>  2 Chronic viral hepatitis                    Infectious Diseas…        1 blue  
#>  3 Encephalitis                               Infectious Diseas…        1 blue  
#>  4 Mycoses                                    Infectious Diseas…        1 blue  
#>  5 HIV                                        Infectious Diseas…        1 blue  
#>  6 Infection of anal and rectal regions       Infectious Diseas…        1 blue  
#>  7 Infection of bones and joints              Infectious Diseas…        1 blue  
#>  8 Infections of the digestive system         Infectious Diseas…        1 blue  
#>  9 Ear and Upper Respiratory Tract Infections Infectious Diseas…        1 blue  
#> 10 Eye infections                             Infectious Diseas…        1 blue  
#> # … with 298 more rows