Map UK Biobank clinical events to phecodes

UK Biobank clinical events sources that are recorded in ICD10 are mapped directly to phecodes, while non-ICD10 sources are mapped to phecodes via ICD10.

Usage

map_clinical_events_to_phecodes2(
  clinical_events,
  all_lkps_maps = NULL,
  min_date_only = FALSE
)

Arguments

clinical_events

A long format data frame created by tidy_clinical_events, tidy_gp_clinical, tidy_gp_scripts or make_clinical_events_db. This can also be a tbl_dbi object.

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a SQLite database containing these tables (see also build_all_lkps_maps() and all_lkps_maps_to_db()). By default, will look for a SQLite datbase called all_lkps_maps.db in the current working directory.

min_date_only

If TRUE, result will be filtered for only the earliest date per eid-phecode pair (date will be recorded as NA for cases where there are no dates).

Value

A data frame with column names 'eid', 'source', 'index', 'code', 'icd10', 'phecode' and 'date'.

Details

Maps the following UK Biobank clinical events sources to phecodes: f40001, f40002, f20002_icd10, f40006, f41270, f40013, f41271, gpc1_r3, gpc2_r3, gpc3_r3, gpc4_r3, gpc1_r2, gpc2_r2, gpc3_r2, gpc4_r2.

Examples