Map codes in a UKB clinical events table

Joins the existing code column to a mapping data frame created by codemapper::get_mapping_df, then filters out any missing values for the new code type. A column is appended containing the newly mapped codes.

Usage

map_codes_ukb_clinical_events(
  clinical_events,
  from,
  to,
  from_colname = NULL,
  to_colname = NULL,
  all_lkps_maps = "all_lkps_maps.db",
  strict_ukb = TRUE
)

Arguments

clinical_events

A clinical events data frame created by ukbwranglr::tidy_clinical_events()

from

Coding type being mapped from

to

Coding type being mapped to

from_colname

Name of column containing codes to be mapped from. If NULL (default), this is assumed to be named 'code'.

to_colname

Name of new column containing mapped codes. If NULL (default), this will equal the value for argument to.

all_lkps_maps

Named list of SQLite database with lookup/mapping tables

strict_ukb

If TRUE, extra checks are performed to check that source and data_coding match ukbwranglr::clinical_events_sources(), and that column names match those expected for a clinical events table.

Value

A clinical events data frame.

Details

The clinical_events data frame can contain multiple sources, but an error is raised by default if these do not all use the same data coding (see ukbwranglr::clinical_events_sources() for recognised data sources and their respective data codings).