Take a sample of rows from non-significant results

May be useful when plotting Manhattan or Q-Q plots for GWAS results to speed up plotting.

Usage

sample_nonsig_results(df, pval_col = "P", sig = 5e-08, n, prop)

Arguments

df

A data frame.

pval_col

Name of column in df containing p values.

sig

Significance threshold. P values equal to or greater than this will be sampled.

n

Provide either n, the number of rows, or prop, the proportion of rows to select. If neither are supplied, n = 1 will be used.

If n is greater than the number of rows in the group (or prop > 1), the result will be silently truncated to the group size. If the proportion of a group size is not an integer, it is rounded down.

prop

Provide either n, the number of rows, or prop, the proportion of rows to select. If neither are supplied, n = 1 will be used.

If n is greater than the number of rows in the group (or prop > 1), the result will be silently truncated to the group size. If the proportion of a group size is not an integer, it is rounded down.

Value

A data frame.

Examples

dummy_manhattan_df() %>%
  sample_nonsig_results(pval_col = "P",
                        sig = 0.001,
                        n = 2)
#> # A tibble: 6 × 5
#>   CHR   SNP    BETA    SE           P
#>   <chr> <chr> <dbl> <dbl>       <dbl>
#> 1 1     b       0    0.02 0.0001     
#> 2 1     c      -0.1  0.03 0.000000001
#> 3 11    l      -0.1  0.03 0.0001     
#> 4 11    o      -0.1  0.03 0.0001     
#> 5 1     a       0.1  0.01 0.05       
#> 6 11    j       0.1  0.01 0.07