Create a UKB data frame with a unique ID column

A convenience function that returns a data frame for a main UK Biobank dataset with a unique ID column, created by concatenating values from a selection of variables.

Usage

create_unique_id_df(
  path,
  delim = "\t",
  ukb_data_dict = ukbwranglr::get_ukb_data_dict(),
  ukb_codings = ukbwranglr::get_ukb_codings(),
  descriptive_colnames = TRUE,
  labelled = TRUE,
  field_ids = c("31", "52", "34", "21000", "53", "96", "50"),
  id_col = "..unique_id",
  remove = TRUE,
  .ignore_duplicate_ids = FALSE
)

Arguments

path

The path to a UK Biobank main dataset file.

delim

Delimiter for the UK Biobank main dataset file. Default is \t

ukb_data_dict

The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ukb_codings

The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

descriptive_colnames

If TRUE, rename columns with longer descriptive names derived from the UK Biobank's data dictionary 'Field' column.

labelled

If TRUE, apply variable and value labels using labelled.

field_ids

A character vector of fields IDs that will be used to create the new unique ID column. These should match the values under column 'Field' in the UK Biobank data dictionary.

id_col

Name of the the new column to be created.

remove

If TRUE, remove input columns from output data frame.

.ignore_duplicate_ids

If TRUE, allow duplicate ID values and raise a warning if any are found. May be helpful for debugging. By default this is FALSE.

Value

A data frame

See also